requirements.txt
cachetools==5.4.0
chardet==5.2.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
filelock==3.15.4
fonttools==4.53.1
kiwisolver==1.4.5
Levenshtein==0.25.1
matplotlib==3.9.2
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pyparsing==3.1.2
pyproject-api==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
rapidfuzz==3.9.6
six==1.16.0
tomli==2.0.1
tox==4.17.1
virtualenv==20.26.3
